Title: 사각형 채우기 3, Time: 134ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 250121/사각형 채우기 3/rectangle-fill-3.java
@@ -2,7 +2,7 @@ import java.util.Scanner;
 
 public class Main {
     public static final int MAX_N = 1000;
-    public static int[] dp = new int[MAX_N + 1];
+    public static long[] dp = new long[MAX_N + 1];
 
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
